# A03_OpenAPI_2206081925.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Marketplace API",
    "description": "Dokumentasi API untuk arsitektur microservices"
  },
  "host": "localhost:8000",
  "basePath": "/api",
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "paths": {
...
    "/wishlist": {
      "get": {
        "summary": "Dapatkan seluruh wishlist milik pengguna",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Daftar wishlist"
          }
        }
      },
      "post": {
        "summary": "Buat wishlist baru",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["name"],
              "properties": {
                "name": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Wishlist berhasil dibuat"
          }
        }
      }
    },
    "/wishlist/{id}": {
      "get": {
        "summary": "Dapatkan item wishlist berdasarkan ID",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Item wishlist ditemukan"
          }
        }
      },
      "put": {
        "summary": "Perbarui wishlist berdasarkan ID",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Wishlist diperbarui"
          }
        }
      },
      "delete": {
        "summary": "Hapus wishlist berdasarkan ID",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Wishlist dihapus"
          }
        }
      }
    },
    "/wishlist/add": {
      "post": {
        "summary": "Tambahkan produk ke wishlist",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["wishlist_id", "product_id"],
              "properties": {
                "wishlist_id": {
                  "type": "string"
                },
                "product_id": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Produk ditambahkan ke wishlist"
          }
        }
      }
    },
    "/wishlist/remove": {
      "delete": {
        "summary": "Hapus produk dari wishlist",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["id"],
              "properties": {
                "id": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Item wishlist dihapus"
          }
        }
      }
    },
...
  }
}